Made RelStorage an extra dependency, along with psycopg2.
Moved the relstorage resolvers in another file to avoid import
problems.
